Move GeoCollect to a separate application
