+++ mapcomposer/app/applications/geocollect/buildjs.cfg
@@ -1,0 +1,17 @@
+[geocollect.js]
+root = static/script
+include =
+    plugins/GeoStoreMissionResourceEditor.js
+    widgets/geoCollect/GcMobileWidgetPanel.js
+    widgets/geoCollect/GeoStoreMissionResourceForm.js
+    widgets/geoCollect/GcDbResourceEditor.js
+    widgets/geoCollect/GcMListResourceEditor.js
+    widgets/geoCollect/GcMobileResourceEditor.js
+    widgets/geoCollect/GcFormResourceEditor.js
+    widgets/geoCollect/GcFormSegResourceEditor.js
+    widgets/geoCollect/GcPViewResourceEditor.js
+    widgets/geoCollect/GcResourceEditor.js
+    widgets/geoCollect/FilterField.js
+    widgets/geoCollect/ComparisonComboBox.js
+    Ext.ux/Spinner.js
+    Ext.ux/SpinnerField.js
