+++ mapcomposer/app/applications/geocollect/static/config/geoCollectManager.js
@@ -4,7 +4,7 @@
    "start":0,
    "limit":20,
    "msmTimeout":30000,
-   "adminUrl": "http://84.33.2.28/opensdi2-manager/",
+   "adminUrl": "http://geocollect.geo-solutions.it/opensdi2-manager/",
    "twitter":{
       "via":"geosolutions_it",
       "hashtags":""
@@ -25,7 +25,7 @@
       ],
       [
          "fr",
-         "Fran̤ais"
+         "Français"
       ],
       [
          "de",
@@ -33,7 +33,7 @@
       ],
       [
          "es",
-         "Espa̱ol"
+         "Español"
       ]
    ],
    "tools":[{
@@ -56,14 +56,18 @@
           "target": "north.tbar",
           "index": 7
         }
-    }
-    
-    
-    ],
+    }],
    "adminTools":[{
         "ptype": "mxp_mapmanager",
         "loginManager": "loginTool",
         "actionTarget": null
+    },{
+        "ptype": "mxp_templatemanager",
+        "loginManager": "loginTool",
+		"actionTarget":{
+          "target": "north.tbar",
+          "index": 0
+        }
     },{
         "ptype": "mxp_myaccount",
         "loginManager": "loginTool",
@@ -79,31 +83,47 @@
           "index": 2
         }
     },{ 
+        "ptype": "mxp_geostore_resource_editor",
+        "category": "GEOCOLLECT",
+        "buttonText": "Mission Templates",
+        "actionTarget":{
+             "target": "north.tbar",
+             "index": 3
+            }
+
+    },{ 
+        "ptype": "mxp_geostore_resource_editor",
+        "category": "MAP",
+        "buttonText": "Maps",
+        "actionTarget":{
+             "target": "north.tbar",
+             "index": 5
+            },
+		"attributeFields":[{
+			"xtype":"textfield",
+			"id":"attribute.owner",
+			"anchor":'95%',
+			"fieldLabel": "Owner",
+			"name":"attribute.owner"
+		}]
+
+    },{ 
         "ptype": "mxp_geostore_mission_resource_editor",
         "category": "GEOCOLLECT",
          "loginManager": "loginTool",
         "buttonText": "Mission Configuration",
              "actionTarget":{
              "target": "north.tbar",
-             "index": 3
+         "index": 6
             },
-            "attributeFields":[{
-  "xtype":"textfield",
-  "id":"attribute.owner",
-  "anchor":'95%',
-  "fieldLabel": "Owner",
-  "name":"attribute.owner"
- }],
-
            "resourceEditor":{
 			"xtype":"mxp_gc_resource_editor",
 			"ref":"/missionResEdit",
 			"authParam":"authkey"
-			},
-           
-
+		},
     },{ 
-        "ptype": "mxp_filebrowser",
+        "ptype": "mxp_servicemanager",
+	"buttonText": "Photos",
         "notDuplicateOutputs":true,
         "actionTarget":{
              "target": "north.tbar",
@@ -135,8 +155,7 @@
           "target": "north.tbar",
           "index": 1 
         }
-    },
-    {
+    },{
         "ptype": "mxp_login",
         "pluginId": "loginTool",
         "actionTarget":{
