+++ mapcomposer/app/applications/geocollect/static/script/Ext.ux/Spinner.js
