+++ mapcomposer/app/applications/geocollect/static/script/Ext.ux/SpinnerField.js
