+++ mapcomposer/app/applications/geocollect/static/script/widgets/geoCollect/ComparisonComboBox.js
