+++ mapcomposer/app/applications/geocollect/static/script/widgets/geoCollect/FilterField.js
