+++ mapcomposer/app/applications/geocollect/static/script/widgets/geoCollect/GcDbResourceEditor.js
