+++ mapcomposer/app/applications/geocollect/static/script/widgets/geoCollect/GcMListResourceEditor.js
