+++ mapcomposer/app/applications/geocollect/static/script/widgets/geoCollect/MobileXDataTypeComboBox.js
