+++ mapcomposer/app/applications/geocollect/static/script/widgets/geoCollect/MobileXtypeComboBox.js
