+++ mapcomposer/app/applications/geocollect/static/script/widgets/geoCollect/OptionsCreator.js
