+++ mapcomposer/app/applications/geocollect/static/script/widgets/geoCollect/xtype/XtypeActionSend.js
