+++ mapcomposer/app/applications/geocollect/static/script/widgets/geoCollect/xtype/XtypeLabel.js
