+++ mapcomposer/app/applications/geocollect/static/script/widgets/geoCollect/xtype/XtypeMap.js
