+++ mapcomposer/app/applications/geocollect/static/script/widgets/geoCollect/xtype/XtypeSepIc.js
