+++ mapcomposer/app/applications/geocollect/templates/base.html
@@ -1,0 +1,21 @@
+<!DOCTYPE html>
+<html>
+    <head>
+        <!--meta http-equiv="X-UA-Compatible" content="IE=8" /-->
+        <title>MapStore</title>
+        <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
+        <link rel="shortcut icon" href="theme/app/img/favicon.ico">
+		
+        <!-- Ext resources -->
+        <link rel="stylesheet" type="text/css" href="externals/ext/resources/css/ext-all.css">
+        <link rel="stylesheet" type="text/css" href="externals/ext/resources/css/xtheme-gray.css">
+        <script type="text/javascript" src="externals/ext/adapter/ext/ext-base.js"></script>
+        <script type="text/javascript" src="externals/ext/ext-all.js"></script>
+		
+        <% render extrahead %>
+    </head>
+    <body>
+        <% render content %>
+		<div style="visibility:hidden"><canvas id="printcanvas" /></div>
+    </body>
+</html>
