--- mapcomposer/app/static/config/geocollect.js
@@ -1,282 +1,0 @@
-{
-   "scaleOverlayMode": "basic",
-   "gsSources":{ 
-		"mapquest": {
-			"ptype": "gxp_mapquestsource"
-		}, 
-		"osm": { 
-			"ptype": "gxp_osmsource"
-		},
-		"google": {
-			"ptype": "gxp_googlesource" 
-		},
-		"bing": {
-			"ptype": "gxp_bingsource" 
-		}, 
-		"ol": { 
-			"ptype": "gxp_olsource" 
-		},
-		"geoserver": {
-			"url":"http://geocollect.geo-solutions.it/geoserver/wms",
-			"ptype":"gxp_wmssource"
-		}
-	},
-	"loadingPanel": {
-		"width": 100,
-		"height": 100,
-		"center": true
-	},
-	"map": {
-		"projection": "EPSG:900913",
-		"units": "m",
-		"center": [
-            		988537.00295749,
-            		5532384.1329431
-        	],
-        	"zoom": 12,
-        	"maxExtent": [
-        	    -20037508.34,
-       	     -20037508.34,
-       	     20037508.34,
-      	 	     20037508.34
-      	  	],
-		"layers": [
-			{
-				"source": "google",
-				"title": "Google Roadmap",
-				"name": "ROADMAP",
-				"visibility": false,
-         			"group": "background"
-			},{
-				"source": "google",
-				"title": "Google Terrain",
-				"name": "TERRAIN",
-				"visibility": false,
-         			"group": "background"
-			},{
-				"source": "google",
-				"title": "Google Hybrid",
-				"name": "HYBRID",
-				"visibility": false,
-         			"group": "background"
-			},{
-				"source": "mapquest",
-				"title": "MapQuest OpenStreetMap",
-				"name": "osm",
-				"visibility": false,
-         			"group": "background"
-			},{
-				"source": "osm",
-				"title": "Open Street Map",
-                		"visibility": true,
-				"name": "mapnik",
-				"group": "background"
-			},{
-				"source": "bing",
-				"title": "Bing Aerial",
-				"name": "Aerial",
-				"visibility": false,
-         			"group": "background"
-			},{
-				"source": "bing",
-				"title": "Bing Aerial With Labels",
-				"name": "AerialWithLabels",
-				"visibility": false,
-         			"group": "background"
-			},{
-				"source": "ol",
-				"group": "background",
-				"fixed": true,
-				"type": "OpenLayers.Layer",
-				"visibility": false,
-				"args": [
-					"None", {"visibility": false}
-				]
-			},{
-				"source": "geoserver",
-				"infoFormat": "application/vnd.ogc.gml",
-				"name": "it.geosolutions:latest_rilievi",
-				"title": "Rilevamenti Effettuati",
-				"visibility": true,
-				"opacity": 1,
-				"selected": false,
-				"format": "image/png",
-				"styles": "rilievi",
-				"transparent": true
-			},{
-				"source": "geoserver",
-				"name": "it.geosolutions:punti_abbandono",
-				"title": "Punti Abbandono",
-				"visibility": true,
-				"opacity": 1,
-				"selected": false,
-				"format": "image/png",
-				"styles": "abbandono",
-				"transparent": true
-			}
-
-		]
-	},
-	"scaleOverlayUnits":{
-        "bottomOutUnits":"mi",    
-        "bottomInUnits":"ft",    
-        "topInUnits":"m",    
-        "topOutUnits":"km"
-    },
-	"customTools":[
-		{
-			"ptype": "gxp_embedmapdialog",
-			"actionTarget": {"target": "paneltbar", "index": 2},
-			"embeddedTemplateName": "viewer",
-			"showDirectURL": true
-		}, {
-			"ptype": "gxp_categoryinitializer",
-            "silentErrors": true,
-			"neededCategories": ["MAPSTORECONFIG", "GEOCOLLECT", "MAP"]
-		}, {
-		   "ptype": "gxp_mouseposition",
-		   "displayProjectionCode":"EPSG:4326",
-		   "customCss": "font-weight: bold; text-shadow: 1px 0px 0px #FAFAFA, 1px 1px 0px #FAFAFA, 0px 1px 0px #FAFAFA,-1px 1px 0px #FAFAFA, -1px 0px 0px #FAFAFA, -1px -1px 0px #FAFAFA, 0px -1px 0px #FAFAFA, 1px -1px 0px #FAFAFA, 1px 4px 5px #aeaeae;color:#050505 "
-		}, {
-			"ptype": "gxp_addlayer",
-			"showCapabilitiesGrid": true,
-			"useEvents": false,
-			"showReport": "never",
-			"directAddLayer": false,
-			"id": "addlayer"
-		}, {
-			"actions": ["-"], 
-			"actionTarget": "paneltbar"
-		}, {
-			"ptype": "gxp_geolocationmenu",
-			"actionTarget": {"target": "paneltbar", "index": 23},
-			"toggleGroup": "toolGroup"
-		}, {
-			"actions": ["->"], 
-			"actionTarget": "paneltbar"
-		}, {
-			"ptype": "gxp_help",
-			"actionTarget": "paneltbar",
-			"text": "Help",
-			"tooltip":"MapStore Guide",
-			"index": 24,
-			"showOnStartup": false,
-			"fileDocURL": "MapStore-Help.pdf"
-        }, {
-			"ptype": "gxp_about",
-			"poweredbyURL": "http://www.geo-solutions.it/about/contacts/",
-			"actionTarget": {"target": "panelbbar", "index": 1}
-		}, {
-			"ptype": "gxp_languageselector",
-			"actionTarget": {"target": "panelbbar", "index": 3}
-		}, {
-			"ptype": "gxp_wmsgetfeatureinfo_menu", 
-			"regex": "[\\s\\S]*[\\w]+[\\s\\S]*",
-			"useTabPanel": true,
-			"toggleGroup": "toolGroup",
-			"actionTarget": {"target": "paneltbar", "index": 20}
-		}
-	],
-	"tools": [
-		{
-			"ptype": "gxp_layertree",
-			"outputConfig": {
-				"id": "layertree"
-			},
-			"outputTarget": "tree",
-			"localIndexs":{
-					"it": 0,
-					"de": 1,
-					"en": 2,
-					"fr": 3
-			}
-		}, {
-			"ptype": "gxp_legend",
-			"outputTarget": "legend",
-			"outputConfig": {
-				"autoScroll": true
-			},
-			"legendConfig" : {
-				"legendPanelId" : "legendPanel",
-				"defaults": {
-					"style": "padding:5px",                  
-					"baseParams": {
-						"FORMAT": "image/jpeg",
-						"LEGEND_OPTIONS": "forceLabels:on;fontSize:10",
-						"WIDTH": 20, "HEIGHT": 20
-					}
-				}
-			}
-		}, {
-			"ptype": "gxp_addlayers",
-			"actionTarget": "tree.tbar",
-			"id": "addlayers"
-		}, {
-			"ptype": "gxp_removelayer",
-			"actionTarget": ["tree.tbar", "layertree.contextMenu"]
-		}, {
-			"ptype": "gxp_removeoverlays",
-			"actionTarget": "tree.tbar"
-		}, {
-			"ptype": "gxp_addgroup",
-			"actionTarget": "tree.tbar"
-		}, {
-			"ptype": "gxp_removegroup",
-			"actionTarget": ["tree.tbar", "layertree.contextMenu"]
-		}, {
-			"ptype": "gxp_groupproperties",
-			"actionTarget": ["tree.tbar", "layertree.contextMenu"]
-		}, {
-			"ptype": "gxp_layerproperties",
-			"actionTarget": ["tree.tbar", "layertree.contextMenu"]
-		}, {
-			"ptype": "gxp_zoomtolayerextent",
-			"actionTarget": {"target": "layertree.contextMenu", "index": 0}
-		}, {
-			"ptype":"gxp_geonetworksearch",
-			"actionTarget": ["layertree.contextMenu"]
-		}, {
-			"ptype": "gxp_zoomtoextent",
-			"actionTarget": {"target": "paneltbar", "index": 15}
-		}, {
-			"ptype": "gxp_navigation", "toggleGroup": "toolGroup",
-			"actionTarget": {"target": "paneltbar", "index": 16}
-		}, {
-			"actions": ["-"], "actionTarget": "paneltbar"
-		}, {
-			"ptype": "gxp_zoombox", "toggleGroup": "toolGroup",
-			"actionTarget": {"target": "paneltbar", "index": 17}
-		}, {
-			"ptype": "gxp_zoom",
-			"actionTarget": {"target": "paneltbar", "index": 18}
-		}, {
-			"actions": ["-"], "actionTarget": "paneltbar"
-		}, {
-			"ptype": "gxp_navigationhistory",
-			"actionTarget": {"target": "paneltbar", "index": 19}
-		}, {
-			"actions": ["-"], "actionTarget": "paneltbar"
-		}, {
-			"ptype": "gxp_wmsgetfeatureinfo_menu", 
-			"regex": "[\\s\\S]*[\\w]+[\\s\\S]*",
-			"picturesBrowserConfig": {
-				"baseUrl": "http://geocollect.geo-solutions.it/opensdi2-manager/mvc/fileManager/extJSbrowser",
-				"folder": "/geocollect/media/punti_abbandono/",
-				"featureProperty": "MY_ORIG_ID",
-				"urlSuffix":"/2"
-			},
-			"useTabPanel": true,
-			"toggleGroup": "toolGroup",
-			"actionTarget": {"target": "paneltbar", "index": 20}
-		}, {
-			"ptype": "gxp_addlayer",
-			"showCapabilitiesGrid": false,
-			"id": "addlayer"
-		}, {
-			"actions": ["-"], "actionTarget": "paneltbar"
-		}, {
-			"ptype": "gxp_measure", "toggleGroup": "toolGroup",
-			"actionTarget": {"target": "paneltbar", "index": 21}
-		}
-	]
-}
