+++ mapcomposer/buildjs.cfg
@@ -420,20 +420,6 @@ include =
     src/mxp/plugins/EntityManager.js
     src/mxp/plugins/StaticPage.js
     src/mxp/plugins/HelpButton.js
-    src/mxp/plugins/GeoStoreMissionResourceEditor.js
-    src/mxp/widgets/geoCollect/GcMobileWidgetPanel.js
-    src/mxp/widgets/geoCollect/GeoStoreMissionResourceForm.js
-    src/mxp/widgets/geoCollect/GcDbResourceEditor.js
-    src/mxp/widgets/geoCollect/GcMListResourceEditor.js
-    src/mxp/widgets/geoCollect/GcMobileResourceEditor.js
-    src/mxp/widgets/geoCollect/GcFormResourceEditor.js
-    src/mxp/widgets/geoCollect/GcFormSegResourceEditor.js
-    src/mxp/widgets/geoCollect/GcPViewResourceEditor.js
-    src/mxp/widgets/geoCollect/GcResourceEditor.js
-    src/mxp/widgets/geoCollect/FilterField.js
-    src/mxp/widgets/geoCollect/ComparisonComboBox.js
-	src/Ext.ux/Spinner.js
-	src/Ext.ux/SpinnerField.js
 	
 exclude = 
 	src/config.js
